requirements.txt
Flask == 2.0.3
gunicorn == 20.0.4
wfastcgi == 3.0.0
Flask
Flask-Login
psycopg2-binary==2.9.5

